[peerdist] Fix segment identifier constant on big-endian targets
The "MS_P2P_CACHING" constant (used as part of the HMAC digest
calculation for the segment identifier) is a UTF-16LE string.  On a
big-endian target, a wide-character string literal will have the wrong
endianness.

Fix by using a byte array rather than a wide-character string.

+++ src/net/pccrc.c
@@ -106,7 +106,7 @@ static void peerdist_info_segment_hash ( struct peerdist_info_segment *segment,
 	struct digest_algorithm *digest = info->digest;
 	uint8_t ctx[ hmac_ctxsize ( digest ) ];
 	size_t digestsize = info->digestsize;
-	static const uint16_t magic[] = PEERDIST_SEGMENT_ID_MAGIC;
+	static const uint8_t magic[] = PEERDIST_SEGMENT_ID_MAGIC;
 
 	/* Sanity check */
 	assert ( digestsize <= sizeof ( segment->hash ) );
